Assorted improvements to smooth_mode

+++ src/commands.rs
@@ -1,6 +1,6 @@
 use std::process::{Command, Stdio};
 
-pub fn get_gpu_temp() -> u32 {
+pub fn get_gpu_temp() -> u64 {
     let output = Command::new("nvidia-smi")
         .args(&["--query-gpu=temperature.gpu", "--format=csv,noheader"])
         .output()
@@ -10,12 +10,12 @@ pub fn get_gpu_temp() -> u32 {
 
     temp_str
         .trim()
-        .parse::<u32>()
-        .unwrap_or(0 as u32)
+        .parse::<u64>()
+        .unwrap_or(0 as u64)
         .clamp(0, 200)
 }
 
-pub fn get_fan_speed() -> u32 {
+pub fn get_fan_speed() -> u64 {
     let output = Command::new("nvidia-smi")
         .args(&["--query-gpu=fan.speed", "--format=csv,noheader"])
         .output()
@@ -24,7 +24,7 @@ pub fn get_fan_speed() -> u32 {
     let _speed_str = String::from_utf8_lossy(&output.stdout);
     let speed_str = _speed_str.trim().replace(" %", "");
 
-    speed_str.parse::<u32>().unwrap_or(0 as u32).clamp(0, 100)
+    speed_str.parse::<u64>().unwrap_or(0 as u64).clamp(0, 100)
 }
 
 pub fn set_fan_control(mode: u8) -> Result<(), Box<dyn std::error::Error>> {
@@ -50,7 +50,7 @@ pub fn set_fan_control(mode: u8) -> Result<(), Box<dyn std::error::Error>> {
     }
 }
 
-pub fn set_fan_speed(speed: u32) -> Result<(), Box<dyn std::error::Error>> {
+pub fn set_fan_speed(speed: u64) -> Result<(), Box<dyn std::error::Error>> {
     let output = Command::new("sudo")
         .args(&[
             "nvidia-settings",
